# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: WeatherEvent Mashup API
  description: |
    ## 🌤️ WeatherEvent Mashup
    
    Une API qui combine :
    - **Météo** (OpenWeatherMap)
    - **Événements culturels** (Ticketmaster)
    - **Recommandations personnalisées** (Service interne)
    
    ### Concepts démontrés :
    - Appels parallèles avec CompletableFuture.allOf()
    - Cache Redis
    - Circuit Breaker Resilience4j
  version: 1.0.0
  contact:
    name: Malaïka Ladéesse Assondji
    email: middleware@univ-cm

servers:
  - url: http://localhost:8080/api/v1
    description: Serveur de développement local

tags:
  - name: Agenda
    description: 📋 Endpoints principaux - Mashup météo + événements + recommandations
  - name: Benchmark
    description: 📊 Comparaison des performances (séquentiel vs parallèle)
  - name: Health
    description: ❤️ Monitoring et santé du service

paths:
  /agenda:
    get:
      tags:
        - Agenda
      summary: Obtenir l'agenda complet
      description: |
        Combine en **appels parallèles** :
        - Météo pour la ville
        - Événements culturels
        - Recommandations personnalisées
        
        ⚡ **Performance** : Temps = max(latences) au lieu de somme
      operationId: getAgenda
      parameters:
        - name: city
          in: query
          required: true
          description: Nom de la ville
          schema:
            type: string
            minLength: 2
            maxLength: 100
            pattern: "^[a-zA-Z\\s-]+$"
          example: Paris
        - name: date
          in: query
          required: true
          description: Date au format YYYY-MM-DD
          schema:
            type: string
            format: date
            pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          example: 2024-12-25
      responses:
        '200':
          description: Agenda construit avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgendaResponse'
              example:
                city: Paris
                date: 2024-12-25
                weather:
                  city: Paris
                  condition: Sunny
                  temperature: 22.5
                  feelsLike: 23.0
                  humidity: 65
                  description: Pleasant weather
                  fallback: false
                events:
                  - id: "1"
                    name: Jazz Festival
                    venue: Downtown Theater
                    city: Paris
                    category: Music
                recommendations:
                  - id: "1"
                    activity: Visit Historic Center
                    venue: Paris Old Town
                    reason: Beautiful architecture
                    priority: 1
                    indoor: false
                processingTimeMs: 312
                mode: PARALLEL
        '400':
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ville non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /agenda/benchmark:
    get:
      tags:
        - Benchmark
      summary: Comparer les performances
      description: |
        Compare le temps d'exécution entre :
        - **Mode Séquentiel** : somme des latences
        - **Mode Parallèle** : maximum des latences
        
        📊 **Gain attendu** : 2x à 3x plus rapide
      operationId: benchmarkAgenda
      parameters:
        - name: city
          in: query
          required: true
          schema:
            type: string
          example: Paris
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: 2024-12-25
      responses:
        '200':
          description: Benchmark complété
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenchmarkResult'
              example:
                sequentialTimeMs: 550
                parallelTimeMs: 200
                speedupFactor: 2.75

  /health:
    get:
      tags:
        - Health
      summary: Vérification de santé
      description: Retourne l'état du service (UP/DOWN/DEGRADED)
      operationId: healthCheck
      responses:
        '200':
          description: Service en bonne santé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /health/circuit-breakers:
    get:
      tags:
        - Health
      summary: État des circuit breakers
      description: Affiche l'état de chaque circuit breaker (CLOSED/OPEN/HALF_OPEN)
      operationId: getCircuitBreakerStatus
      responses:
        '200':
          description: États des circuit breakers
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  enum: [CLOSED, OPEN, HALF_OPEN]
              example:
                weatherApi: CLOSED
                eventsApi: CLOSED
                recommendationsApi: CLOSED

components:
  schemas:
    BenchmarkStats:
      type: object
      properties:
        mean:
          type: number
          format: double
          description: Moyenne en millisecondes
        stdDev:
          type: number
          format: double
          description: Écart-type
        percentile95:
          type: integer
          format: int64
          description: Percentile 95
        min:
          type: integer
          format: int64
          description: Temps minimum
        max:
          type: integer
          format: int64
          description: Temps maximum
        count:
          type: integer
          description: Nombre d'itérations

    RigorousBenchmarkResult:
        type: object
        properties:
          city:
            type: string
          date:
            type: string
          iterations:
            type: integer
          warmup:
            type: integer
          sequentialStats:
            $ref: '#/components/schemas/BenchmarkStats'
          parallelStats:
            $ref: '#/components/schemas/BenchmarkStats'
          speedupMean:
            type: number
            format: double
            description: Facteur d'accélération moyen
    ModeEnum:
      type: string
      enum:
        - PARALLEL
        - SEQUENTIAL
      description: Mode d'exécution

    StatusEnum:
      type: string
      enum:
        - UP
        - DOWN
        - DEGRADED
      description: Statut du service

    WeatherResponse:
      type: object
      properties:
        city:
          type: string
          description: Nom de la ville
        condition:
          type: string
          description: Condition météo (Sunny, Rainy, etc.)
        temperature:
          type: number
          format: double
          description: Température en degrés Celsius
        feelsLike:
          type: number
          format: double
          description: Température ressentie
        humidity:
          type: integer
          description: Humidité en pourcentage
        description:
          type: string
          description: Description textuelle de la météo
        fallback:
          type: boolean
          description: Indique si c'est une réponse de secours
        cachedAt:
          type: string
          format: date-time
          description: Date de mise en cache
        windSpeed:
          type: number
          format: double
          description: Vitesse du vent en km/h

    EventResponse:
      type: object
      properties:
        id:
          type: string
          description: Identifiant unique
        name:
          type: string
          description: Nom de l'événement
        venue:
          type: string
          description: Lieu de l'événement
        city:
          type: string
          description: Ville
        category:
          type: string
          description: Catégorie (Music, Art, etc.)
        fallback:
          type: boolean
          description: Indique si c'est une réponse de secours

    RecommendationResponse:
      type: object
      properties:
        id:
          type: string
          description: Identifiant unique
        activity:
          type: string
          description: Activité recommandée
        venue:
          type: string
          description: Lieu recommandé
        reason:
          type: string
          description: Raison de la recommandation
        priority:
          type: integer
          minimum: 1
          maximum: 5
          description: Priorité (1 = plus recommandé)
        indoor:
          type: boolean
          description: Activité en intérieur ?

    ApiStatus:
      type: object
      properties:
        weatherApiAvailable:
          type: boolean
          description: API météo disponible ?
        eventsApiAvailable:
          type: boolean
          description: API événements disponible ?
        recommendationsApiAvailable:
          type: boolean
          description: Service recommandations disponible ?

    AgendaResponse:
      type: object
      properties:
        city:
          type: string
        date:
          type: string
        weather:
          $ref: '#/components/schemas/WeatherResponse'
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventResponse'
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/RecommendationResponse'
        processingTimeMs:
          type: integer
          format: int64
          description: Temps d'exécution en millisecondes
        mode:
          $ref: '#/components/schemas/ModeEnum'
        apiStatus:
          $ref: '#/components/schemas/ApiStatus'

    BenchmarkResult:
      type: object
      properties:
        sequentialTimeMs:
          type: integer
          format: int64
          description: Temps en mode séquentiel (ms)
        parallelTimeMs:
          type: integer
          format: int64
          description: Temps en mode parallèle (ms)
        speedupFactor:
          type: number
          format: double
          description: Facteur d'accélération (seq/par)
        sequentialResponse:
          $ref: '#/components/schemas/AgendaResponse'
        parallelResponse:
          $ref: '#/components/schemas/AgendaResponse'

    HealthResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/StatusEnum'
        service:
          type: string
          description: Nom du service
        version:
          type: string
          description: Version de l'API
        timestamp:
          type: string
          format: date-time
          description: Horodatage

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          description: Code HTTP
        error:
          type: string
          description: Type d'erreur
        message:
          type: string
          description: Message détaillé
        path:
          type: string
          description: Chemin de la requête
        requestId:
          type: string
          description: Identifiant de requête